+++ spec/preview-element-spec.js
@@ -42,50 +42,5 @@ describe('preview-element', function () {
     it('should have a path', function () {
       expect(preview.getPath()).toEqual('/test/path.txt')
     })
-
-    describe('when element is attached to DOM', function () {
-      beforeEach(function () {
-        spyOn(atom.workspace, 'open')
-        jasmine.attachToDOM(preview)
-      })
-
-      describe('when clicked', function () {
-        beforeEach(function () {
-          preview.click()
-        })
-
-        it('should open current path', function () {
-          expect(atom.workspace.open).toHaveBeenCalledWith('/test/path.txt')
-        })
-
-        describe('when cleared', function () {
-          beforeEach(function () {
-            preview.clear()
-          })
-
-          describe('when clicked', function () {
-            beforeEach(function () {
-              atom.workspace.open.reset()
-              preview.click()
-            })
-
-            it('should not open anything', function () {
-              expect(atom.workspace.open).not.toHaveBeenCalled()
-            })
-          })
-        })
-
-        describe('when detached from DOM', function () {
-          beforeEach(function () {
-            atom.workspace.open.reset()
-            preview.remove()
-          })
-
-          it('should should no longer open anything when clicked', function () {
-            expect(atom.workspace.open).not.toHaveBeenCalled()
-          })
-        })
-      })
-    })
   })
 })